# docs/swagger.yaml
basePath: /api/v1
definitions:
  entity.ArticleDetail:
    properties:
      author:
        $ref: '#/definitions/entity.ProfilePreview'
      body:
        type: string
      createdAt:
        type: string
      description:
        type: string
      favorited:
        type: boolean
      favoritesCount:
        type: integer
      slug:
        type: string
      tagList:
        items:
          type: string
        type: array
      title:
        type: string
      updatedAt:
        type: string
    type: object
  entity.ArticlePreview:
    properties:
      author:
        $ref: '#/definitions/entity.ProfilePreview'
      createdAt:
        type: string
      description:
        type: string
      favorited:
        type: boolean
      favoritesCount:
        type: integer
      slug:
        type: string
      tagList:
        items:
          type: string
        type: array
      title:
        type: string
      updatedAt:
        type: string
    type: object
...
  entity.ProfilePreview:
    properties:
      bio:
        type: string
      followersCount:
        type: integer
      following:
        type: boolean
      image:
        type: string
      username:
        type: string
    type: object
...
  request.ArticleCreate:
    properties:
      body:
        example: this is article content
        maxLength: 50000
        type: string
      description:
        example: This is description
        maxLength: 255
        type: string
      tagList:
        example:
        - go
        - fiber
        - api
        - clean-arch
        items:
          type: string
        maxItems: 10
        type: array
      title:
        example: This is title - to generate slug
        maxLength: 255
        minLength: 2
        type: string
    required:
    - body
    - description
    - tagList
    - title
    type: object
  request.ArticleCreateRequest:
    properties:
      article:
        $ref: '#/definitions/request.ArticleCreate'
    type: object
  request.ArticleUpdate:
    properties:
      body:
        example: this is article content
        maxLength: 50000
        type: string
      description:
        example: This is description
        maxLength: 255
        type: string
      title:
        example: This is title - to generate slug
        maxLength: 255
        type: string
    type: object
  request.ArticleUpdateRequest:
    properties:
      article:
        $ref: '#/definitions/request.ArticleUpdate'
    type: object
...
  response.ArticleDetailResponse:
    properties:
      article:
        $ref: '#/definitions/entity.ArticleDetail'
    type: object
  response.ArticlePreviewsResponse:
    properties:
      articles:
        items:
          $ref: '#/definitions/entity.ArticlePreview'
        type: array
      articlesCount:
        type: integer
      limit:
        example: 10
        type: integer
      offset:
        example: 0
        type: integer
      total:
        example: 18
        type: integer
    type: object
...
  response.Error:
    properties:
      error:
        example: message
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Realworld API using Golang + Gin + Clean Architecture
  title: Realworld Fiber Clean API
  version: "1.0"
paths:
  /articles:
    get:
      description: Get all articles (filter by author, favorited, tag)
      operationId: articles-get-all
      parameters:
      - description: Limit number of results
        format: int64
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        format: int64
        in: query
        name: offset
        type: integer
      - description: Filter by author username
        in: query
        name: author
        type: string
      - description: Filter by favorited username
        in: query
        name: favorited
        type: string
      - description: Filter by tag
        in: query
        name: tag
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ArticlePreviewsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - BearerAuth: []
      summary: Get all articles
      tags:
      - articles
    post:
      consumes:
      - application/json
      description: Create Article
      operationId: articles-create
      parameters:
      - description: Create Article
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ArticleCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ArticleDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - BearerAuth: []
      summary: Create Article
      tags:
      - articles
  /articles/{slug}:
    delete:
      description: Delete article by slug
      operationId: articles-delete-by-slug
      parameters:
      - description: Article slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - BearerAuth: []
      summary: Delete article
      tags:
      - articles
    get:
      description: Get article by slug
      operationId: articles-get-by-slug
      parameters:
      - description: Article slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ArticleDetailResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - BearerAuth: []
      summary: Get article
      tags:
      - articles
    put:
      description: Put article by slug
      operationId: articles-put-by-slug
      parameters:
      - description: Article slug
        in: path
        name: slug
        required: true
        type: string
      - description: Update Article
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ArticleUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ArticleDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - BearerAuth: []
      summary: Put article
      tags:
      - articles
...
  /articles/feed:
    get:
      description: Get feed articles (from followed authors)
      operationId: articles-get-feed
      parameters:
      - description: Limit number of results
        format: int64
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        format: int64
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ArticlePreviewsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - BearerAuth: []
      summary: Get feed articles
      tags:
      - articles
...
securityDefinitions:
  BearerAuth:
    description: Type "Token" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"